Moved crypto crate up to root src level
